Add files via upload
Goals set less than $20,000 had a higher percentage of success than failure. 
Kickstarter launched in May were the most successful. 
A Kickstarter that exceeded $50,000 was unlikely to be successful.   


December had the least amount of successful campaigns but it also seems to be around 50% so I would like to have another chart that was a percent outcome based of the months to further analysis successful and failed months.
Outcomes based off regions would be a limitation since they are all based off difference currency and currency is not equal to different currency. I would recommend having a new column that would convert all currency into a single currency to weigh it better. 

I could make a new pivot table and chart to show how the length of a Kickstarter affects its outcome.